Adding Graph Function class which takes care of the point 13 of the programming project
The Graph Function class constructor takes a GraphADT as a parameter.

Calling the function retrieveCliques() returns a Iterable<Integer[]> in which each Integer[] is a symbol array that is a clique and their elements. That point was fullfilled using backtracking with Skiena structure. The clique cannot be inside a bigger clique and the cost of this algorithm is exponential, when the clique group expands, the more connections must check to confirm that it's actually a clique.